+++ src/program/ui/InputEditorView.h
@@ -78,7 +78,7 @@ public slots:
 
     void moveAgainSection(int logicalIndex, int oldVisualIndex, int newVisualIndex);
     void scrollToFrame(unsigned long long frame);
-    void holdSection(int logicalIndex);
+    void clickSection(int logicalIndex);
 
 protected:
     void mousePressEvent(QMouseEvent *event) override;